Tehtävät 4.21 - 4.22

+++ osa4/blogilista/controllers/bloglist.js
@@ -41,8 +41,17 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  console.log(user._id)
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+    console.log('Blog removed')
+  } else {
+    return response.status(401).json({ error: 'Not allowed' })
+  }
 })
 
 blogRouter.put('/:id', async(request, response) => {
